--- src/main/java/org/folio/qm/converter/MarcConverter.java
@@ -1,12 +1,0 @@
-package org.folio.qm.converter;
-
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.lang.NonNull;
-
-public interface MarcConverter<S, T, F> extends Converter<S, T> {
-
-  T convert(@NonNull S source);
-
-  F supportedType();
-
-}
