+++ src/main/java/org/folio/qm/converter/field/qm/AbstractFieldItemConverter.java
@@ -1,4 +1,4 @@
-package org.folio.qm.converternew.qm;
+package org.folio.qm.converter.field.qm;
 
 import static org.apache.commons.lang3.StringUtils.EMPTY;
 
@@ -16,7 +16,7 @@ import org.marc4j.marc.Subfield;
 import org.marc4j.marc.VariableField;
 import org.marc4j.marc.impl.DataFieldImpl;
 
-import org.folio.qm.converternew.FieldItemConverter;
+import org.folio.qm.converter.field.FieldItemConverter;
 import org.folio.qm.domain.dto.FieldItem;
 
 public abstract class AbstractFieldItemConverter implements FieldItemConverter {
