+++ src/main/java/org/folio/qm/converter/field/qm/AdditionalCharacteristicsFieldItemConverter.java
@@ -1,4 +1,4 @@
-package org.folio.qm.converternew.qm;
+package org.folio.qm.converter.field.qm;
 
 import static org.folio.qm.converter.elements.AdditionalMaterialConfiguration.resolveByCode;
 import static org.folio.qm.converter.elements.Constants.ADDITIONAL_CHARACTERISTICS_CONTROL_FIELD;
@@ -11,7 +11,7 @@ import org.marc4j.marc.impl.ControlFieldImpl;
 import org.springframework.stereotype.Component;
 
 import org.folio.qm.converter.elements.ControlFieldItem;
-import org.folio.qm.converternew.FieldItemConverter;
+import org.folio.qm.converter.field.FieldItemConverter;
 import org.folio.qm.domain.dto.FieldItem;
 import org.folio.qm.domain.dto.MarcFormat;
