+++ src/main/java/org/folio/qm/converter/field/qm/GeneralInformationHoldingsFieldItemConverter.java
@@ -1,8 +1,8 @@
-package org.folio.qm.converternew.qm;
+package org.folio.qm.converter.field.qm;
 
-import static org.folio.qm.converter.elements.Constants.AUTHORITY_CONTROL_FIELD_ITEMS;
-import static org.folio.qm.converter.elements.Constants.AUTHORITY_GENERAL_INFORMATION_CONTROL_FIELD_LENGTH;
 import static org.folio.qm.converter.elements.Constants.GENERAL_INFORMATION_CONTROL_FIELD;
+import static org.folio.qm.converter.elements.Constants.HOLDINGS_CONTROL_FIELD_ITEMS;
+import static org.folio.qm.converter.elements.Constants.HOLDINGS_GENERAL_INFORMATION_CONTROL_FIELD_LENGTH;
 
 import java.util.Map;
 
@@ -10,12 +10,12 @@ import org.marc4j.marc.VariableField;
 import org.marc4j.marc.impl.ControlFieldImpl;
 import org.springframework.stereotype.Component;
 
-import org.folio.qm.converternew.FieldItemConverter;
+import org.folio.qm.converter.field.FieldItemConverter;
 import org.folio.qm.domain.dto.FieldItem;
 import org.folio.qm.domain.dto.MarcFormat;
 
 @Component
-public class GeneralInformationAuthorityFieldItemConverter implements FieldItemConverter {
+public class GeneralInformationHoldingsFieldItemConverter implements FieldItemConverter {
 
   @Override
   public VariableField convert(FieldItem field) {
@@ -25,11 +25,11 @@ public class GeneralInformationAuthorityFieldItemConverter implements FieldItemC
 
   @Override
   public boolean canProcess(FieldItem field, MarcFormat marcFormat) {
-    return field.getTag().equals(GENERAL_INFORMATION_CONTROL_FIELD) && marcFormat == MarcFormat.AUTHORITY;
+    return field.getTag().equals(GENERAL_INFORMATION_CONTROL_FIELD) && marcFormat == MarcFormat.HOLDINGS;
   }
 
   protected String restoreControlField(Map<String, Object> contentMap) {
-    return restoreFixedLengthField(AUTHORITY_GENERAL_INFORMATION_CONTROL_FIELD_LENGTH, AUTHORITY_CONTROL_FIELD_ITEMS,
+    return restoreFixedLengthField(HOLDINGS_GENERAL_INFORMATION_CONTROL_FIELD_LENGTH, HOLDINGS_CONTROL_FIELD_ITEMS,
       contentMap, 0);
   }
 }
