+++ src/main/java/org/folio/qm/mapper/MarcTypeMapper.java
@@ -5,25 +5,24 @@ import java.util.Map;
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 import org.mapstruct.Mapper;
-import org.mapstruct.NullValueCheckStrategy;
 
 import org.folio.qm.domain.dto.MarcFormat;
-import org.folio.rest.jaxrs.model.ParsedRecordDto;
+import org.folio.qm.domain.dto.ParsedRecordDto;
 
 @Mapper(componentModel = "spring")
 public interface MarcTypeMapper {
 
-  BiMap<ParsedRecordDto.RecordType, MarcFormat> typeMap = HashBiMap.create(Map.of(
-    ParsedRecordDto.RecordType.MARC_BIB, MarcFormat.BIBLIOGRAPHIC,
-    ParsedRecordDto.RecordType.MARC_AUTHORITY, MarcFormat.AUTHORITY,
-    ParsedRecordDto.RecordType.MARC_HOLDING, MarcFormat.HOLDINGS
+  BiMap<ParsedRecordDto.RecordTypeEnum, MarcFormat> typeMap = HashBiMap.create(Map.of(
+    ParsedRecordDto.RecordTypeEnum.BIB, MarcFormat.BIBLIOGRAPHIC,
+    ParsedRecordDto.RecordTypeEnum.AUTHORITY, MarcFormat.AUTHORITY,
+    ParsedRecordDto.RecordTypeEnum.HOLDING, MarcFormat.HOLDINGS
   ));
 
-  default ParsedRecordDto.RecordType toDto(MarcFormat source) {
+  default ParsedRecordDto.RecordTypeEnum toDto(MarcFormat source) {
     return typeMap.inverse().get(source);
   }
 
-  default MarcFormat fromDto(ParsedRecordDto.RecordType source) {
+  default MarcFormat fromDto(ParsedRecordDto.RecordTypeEnum source) {
     return typeMap.get(source);
   }
 }
