+++ src/main/java/org/folio/qm/service/FieldProtectionSetterService.java
@@ -1,0 +1,8 @@
+package org.folio.qm.service;
+
+import org.folio.qm.domain.dto.QuickMarc;
+
+public interface FieldProtectionSetterService {
+
+  QuickMarc applyFieldProtection(QuickMarc record);
+}
