+++ src/main/java/org/folio/qm/service/impl/ChangeManagerServiceImpl.java
@@ -1,0 +1,46 @@
+package org.folio.qm.service.impl;
+
+import java.util.UUID;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import org.folio.qm.client.SRMChangeManagerClient;
+import org.folio.qm.domain.dto.InitJobExecutionsRqDto;
+import org.folio.qm.domain.dto.InitJobExecutionsRsDto;
+import org.folio.qm.domain.dto.ParsedRecordDto;
+import org.folio.qm.domain.dto.ProfileInfo;
+import org.folio.qm.domain.dto.RawRecordsDto;
+import org.folio.qm.service.ChangeManagerService;
+
+@Service
+@RequiredArgsConstructor
+public class ChangeManagerServiceImpl implements ChangeManagerService {
+
+  private final SRMChangeManagerClient srmClient;
+
+  @Override
+  public ParsedRecordDto getParsedRecordByExternalId(String externalId) {
+    return srmClient.getParsedRecordByExternalId(externalId);
+  }
+
+  @Override
+  public void putParsedRecordByInstanceId(UUID id, ParsedRecordDto recordDto) {
+    srmClient.putParsedRecordByInstanceId(id, recordDto);
+  }
+
+  @Override
+  public InitJobExecutionsRsDto postJobExecution(InitJobExecutionsRqDto jobExecutionDto) {
+    return srmClient.postJobExecution(jobExecutionDto);
+  }
+
+  @Override
+  public void putJobProfileByJobExecutionId(UUID jobExecutionId, ProfileInfo jobProfile) {
+    srmClient.putJobProfileByJobExecutionId(jobExecutionId, jobProfile);
+  }
+
+  @Override
+  public void postRawRecordsByJobExecutionId(UUID jobExecutionId, RawRecordsDto rawRecords) {
+    srmClient.postRawRecordsByJobExecutionId(jobExecutionId, rawRecords);
+  }
+}
