+++ src/main/java/org/folio/qm/service/impl/DataImportJobServiceImpl.java
@@ -13,11 +13,11 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import org.folio.qm.client.SRMChangeManagerClient;
 import org.folio.qm.domain.dto.ParsedRecordDto;
 import org.folio.qm.domain.entity.JobProfile;
 import org.folio.qm.domain.entity.JobProfileAction;
 import org.folio.qm.domain.entity.RecordType;
+import org.folio.qm.service.ChangeManagerService;
 import org.folio.qm.service.CreationStatusService;
 import org.folio.qm.service.DataImportJobService;
 import org.folio.qm.service.JobProfileService;
@@ -33,7 +33,7 @@ public class DataImportJobServiceImpl implements DataImportJobService {
     ParsedRecordDto.RecordTypeEnum.HOLDING, RecordType.MARC_HOLDINGS
   );
 
-  private final SRMChangeManagerClient srmClient;
+  private final ChangeManagerService changeManagerService;
   private final CreationStatusService statusService;
   private final JobProfileService jobProfileService;
   private final FolioExecutionContext folioExecutionContext;
@@ -54,23 +54,23 @@ public class DataImportJobServiceImpl implements DataImportJobService {
 
   private UUID initJob(JobProfile jobProfile, UUID userId) {
     var jobExecutionDto = getDefaultJodExecutionDto(userId, jobProfile);
-    var jobExecutionResponse = srmClient.postJobExecution(jobExecutionDto);
+    var jobExecutionResponse = changeManagerService.postJobExecution(jobExecutionDto);
     var jobExecutionId = jobExecutionResponse.getJobExecutions().get(0).getId();
     saveNewStatus(jobExecutionId);
     return jobExecutionId;
   }
 
   private void updateJobProfile(UUID jobExecutionId, JobProfile jobProfile) {
-    srmClient.putJobProfileByJobExecutionId(jobExecutionId, toJobProfileInfo(jobProfile));
+    changeManagerService.putJobProfileByJobExecutionId(jobExecutionId, toJobProfileInfo(jobProfile));
     saveInProgressStatus(jobExecutionId);
   }
 
   private void sendRecord(ParsedRecordDto parsedRecordDto, UUID jobExecutionId) {
-    srmClient.postRawRecordsByJobExecutionId(jobExecutionId, toRawRecordsDto(parsedRecordDto));
+    changeManagerService.postRawRecordsByJobExecutionId(jobExecutionId, toRawRecordsDto(parsedRecordDto));
   }
 
   private void completeImport(UUID jobExecutionId) {
-    srmClient.postRawRecordsByJobExecutionId(jobExecutionId, toLastRawRecordsDto());
+    changeManagerService.postRawRecordsByJobExecutionId(jobExecutionId, toLastRawRecordsDto());
   }
 
   private JobProfile getJobProfile(ParsedRecordDto recordDto, JobProfileAction action) {
