+++ src/test/java/org/folio/qm/validation/leader/CommonLeaderValidationRuleTest.java
@@ -16,7 +16,7 @@ import org.folio.qm.validation.impl.holdings.HoldingsLeaderValidationRule;
 @UnitTest
 class CommonLeaderValidationRuleTest {
 
-  private final LeaderValidationRule rule = new HoldingsLeaderValidationRule();
+  private final HoldingsLeaderValidationRule rule = new HoldingsLeaderValidationRule();
 
   private static final String WRONG_RECORD_LENGTH = "00241cx\\\\a2200109zn\\45000";
   private static final String LEADER_RECORD_LENGTH_WITH_CHARS = "0a241cx\\\\a2200109zn\\4500";
